Commit for saturday

+++ src/main/java/net/syspherice/controller/AccountController.java
@@ -4,31 +4,43 @@ import java.util.List;
 
 import javax.servlet.http.HttpSession;
 
+import net.syspherice.enumeration.AccountEnum;
 import net.syspherice.form.Account;
 import net.syspherice.service.AccountService;
 import net.syspherice.utils.AbsoluteString;
+import net.syspherice.utils.Common;
 import net.syspherice.utils.MenuBuild;
 import net.syspherice.utils.PagedGenericView;
+import net.syspherice.utils.SessionManage;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.SessionAttributes;
 import org.springframework.web.servlet.ModelAndView;
 
 @Controller
 @RequestMapping("/account")
-public class AccountController {
+public class AccountController{
 	@Autowired
 	private AccountService accountService;
-	
+	private SessionManage sessionManage;
 	@RequestMapping("/index")
 	public ModelAndView Index(HttpSession session) {
+		sessionManage = new SessionManage(session);
+		if(!sessionManage.getIsAdmin()){
+			return new ModelAndView("redirect:/");
+		}
 		return this.IndexData(session);
 	}
 	@RequestMapping("/index/{index}")
 	public ModelAndView Index(@PathVariable("index") Integer index, HttpSession session) {
+		sessionManage = new SessionManage(session);
+		if(!sessionManage.getIsAdmin()){
+			return new ModelAndView("redirect:/");
+		}
 		ModelAndView mv = new ModelAndView("account/index");
 		PagedGenericView<Account> ulist = new PagedGenericView<Account>();
 
@@ -50,7 +62,10 @@ public class AccountController {
 
 	@RequestMapping("/delete/{username}")
 	public ModelAndView delete(@PathVariable("username") String username, HttpSession session) {
-
+		sessionManage = new SessionManage(session);
+		if(!sessionManage.getIsAdmin()){
+			return new ModelAndView("redirect:/");
+		}
 		Boolean result = accountService.delete(username);
 		ModelAndView mv = this.Index(session);
 		
@@ -65,9 +80,32 @@ public class AccountController {
 	}
 	@RequestMapping("/update/{username}")
 	public ModelAndView update(@PathVariable("username") String username, HttpSession session) {
+		sessionManage = new SessionManage(session);
+		if(!sessionManage.getIsAdmin()){
+			return new ModelAndView("redirect:/");
+		}
 		Account accountOrigine = accountService.single(username);
 		Account accountUpdate = accountOrigine;
-		accountUpdate.setState(accountOrigine.getState().compareTo("active")==0?"unactive":"active");
+		accountUpdate.setState(accountOrigine.getState().compareTo("active")==0?"inactive":"active");
+		Boolean result = accountService.update(accountOrigine, accountUpdate);
+		ModelAndView mv = this.Index(session);
+		if(result)
+			mv.addObject(AbsoluteString.noticesuccess, true);
+		else
+			mv.addObject(AbsoluteString.noticefail, true);
+		mv.addObject(AbsoluteString.pathBar, "Account Manage Page");
+		mv.addObject(AbsoluteString.adminmenu, MenuBuild.getAdminMenu("account", session));
+		return mv;
+	}
+	@RequestMapping("/updaterole/{username}")
+	public ModelAndView updaterole(@PathVariable("username") String username, HttpSession session) {
+		sessionManage = new SessionManage(session);
+		if(!sessionManage.getIsAdmin()){
+			return new ModelAndView("redirect:/");
+		}
+		Account accountOrigine = accountService.single(username);
+		Account accountUpdate = accountOrigine;
+		accountUpdate.setRole(accountOrigine.getRole().compareTo("user")==0?"admin":"user");
 		Boolean result = accountService.update(accountOrigine, accountUpdate);
 		ModelAndView mv = this.Index(session);
 		if(result)
@@ -93,6 +131,10 @@ public class AccountController {
 	//for user page
 	@RequestMapping(value = "/detail/{username}")
 	public ModelAndView Detail(@PathVariable("username") String username, ModelMap map, HttpSession session) {
+		sessionManage = new SessionManage(session);
+		if(!sessionManage.getIsAdmin()){
+			return new ModelAndView("redirect:/");
+		}
 		ModelAndView mv = new ModelAndView("account/detail");
 		Account account = accountService.single(username);
